// src/app/api/delivery/delivery.type.ts
// Generated by https://quicktype.io

export interface DeliveryJSON {
    status: number;
    body:   Body[];
}

export interface Body {
    id:                  number;
    status:              string;
    owner:               number;
    created_on:          string;
    tipo:                null | string;
    direccion_recogida:  null | string;
    direcciones_entrega: null | string;
    mapa:                null | string;
    condiciones:         string;
    restricciones:       null | string;
    adicionales:         string;
    precio:              string;
    data:                null | string;
    media:               string;
    relation_with_brand: number;
}